x509: Split internal and public functions
Split the internal and public facing encode functions.  This reduces
the burden for reading the code base, and also emphasizes the CertWriter
must use `check_for_truncated` to ensure encodings are valid and not
truncated prior to returning to the user.
